Add tokenizer destroy module and fix memory leak.

+++ src/inc/nfa.h
@@ -3,7 +3,7 @@
 
 #define NFA_LABEL_HASH_SIZE    127
 #define NFA_SZ_UNMATCH         0
-#define NFA_SZ_INVALID         -1
+#define NFA_SZ_INVALID         (uint32)(-1)
 #define NFA_MARK_CHAR          ((char)0xa4)
 
 typedef struct nfa_edge_map    s_nfa_edge_map_t;
